rename duiration chaos into common chaos
